refactor: move components into Refinements folder
